Check EC2 connection before changing path

+++ test/test_utils/sagemaker.py
@@ -327,13 +327,20 @@ def execute_local_tests(image, pytest_cache_params):
                 ec2_conn.run(pytest_command)
                 print(f"Downloading Test reports for image: {image}")
                 ec2_conn.get(ec2_test_report_path, os.path.join("test", f"{job_type}_{tag}_sm_local.xml"))
+    except:
+        print(f"Error {sys.exc_info()[0]} occurred")
+        print("Exiting after clean up")
     finally:
-        with ec2_conn.cd(path):
-            pytest_cache_util.upload_pytest_cache_from_ec2_to_s3(ec2_conn, path, **pytest_cache_params)
-        print(f"Terminating Instances for image: {image}")
-        ec2_utils.terminate_instance(instance_id, region)
-        print(f"Destroying ssh Key_pair for image: {image}")
-        destroy_ssh_keypair(ec2_client, ec2_key_name)
+        if ec2_conn:
+            with ec2_conn.cd(path):
+                pytest_cache_util.upload_pytest_cache_from_ec2_to_s3(ec2_conn, path, **pytest_cache_params)
+        if instance_id:
+            print(f"Terminating Instances for image: {image}")
+            ec2_utils.terminate_instance(instance_id, region)
+            
+        if ec2_client and ec2_key_name:
+            print(f"Destroying ssh Key_pair for image: {image}")
+            destroy_ssh_keypair(ec2_client, ec2_key_name)
         # return None here to prevent errors from multiprocessing.map(). Without this it returns some object by default
         # which is causing "cannot pickle '_thread.lock' object" error
         return None
